+++ kexec/arch/sh/kexec-sh.h
@@ -23,7 +23,7 @@ int netbsd_sh_load(int argc, char **argv, const char *buf, off_t len,
 void netbsd_sh_usage(void);
 
 char *get_append(void);
-void kexec_sh_setup_zero_page(char *zero_page_buf, int zero_page_size,
+void kexec_sh_setup_zero_page(char *zero_page_buf, size_t zero_page_size,
 			      char *cmd_line);
 
 #endif /* KEXEC_SH_H */
